refactor: move dummy file to src/
